feat(firebase-auth): add base firebase strategy
