+++ src/framework/firebase-auth/firebase-auth.module.ts
@@ -8,12 +8,14 @@
 import { NgModule } from '@angular/core';
 import { NbFirebasePasswordStrategy } from './strategies/password/firebase-password.strategy';
 import { NbFirebaseGoogleStrategy } from './strategies/google/firebase-google.strategy';
+import { NbFirebaseFacebookStrategy } from './strategies/facebook/firebase-facebook.strategy';
 
 
 @NgModule({
   providers: [
     NbFirebasePasswordStrategy,
     NbFirebaseGoogleStrategy,
+    NbFirebaseFacebookStrategy,
   ],
 })
 export class FirebaseAuthModule { }
