+++ src/framework/firebase-auth/public_api.ts
@@ -5,8 +5,9 @@
  */
 
 
+export * from	'./strategies/base/firebase-base.strategy';
 export * from './strategies/password/firebase-password.strategy';
 export * from './strategies/password/firebase-password-strategy.options';
+export * from './strategies/base/firebase-identity-provider-strategy.options';
 export * from './strategies/google/firebase-google.strategy';
-export * from './strategies/google/firebase-google-strategy.options';
 export * from './firebase-auth.module';
